requirements.txt
nonebot2>=2.0.0
nonebot-adapter-onebot>=2.0.0
jinja2>=3.0.0
pillow>=8.0.0
playwright>=1.20.0
aiohttp>=3.8.0
python-dotenv>=0.19.0
pyyaml>=5.4.0
sqlalchemy>=1.4.0
python-imgkit>=1.2.2